feat: created typed redux hooks
