+++ src/store/index.ts
@@ -1,10 +1,2 @@
-import { combineReducers, configureStore } from '@reduxjs/toolkit';
-
-const rootReducer = combineReducers({});
-
-export const store = configureStore({
-  reducer: rootReducer,
-});
-
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
+export { store } from './store';
+export { useAppDispatch, useAppSelector } from './hooks';
